+++ src/main/java/com/dashboard/gmelan/controller/IndexPageController.java
@@ -1,4 +1,4 @@
-package com.dashboard.gmelan.index;
+package com.dashboard.gmelan.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
