+++ src/main/java/com/dashboard/gmelan/controller/NoticeController.java
@@ -1,8 +1,7 @@
-package com.dashboard.gmelan.notice.controller;
+package com.dashboard.gmelan.controller;
 
 import com.dashboard.gmelan.notice.entity.Notice;
 import com.dashboard.gmelan.notice.service.NoticeService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
